+++ react_components/PaginationBoxView.js
@@ -4,6 +4,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import PageView from './PageView';
 import BreakView from './BreakView';
+import { classNameIfDefined } from './utils';
 
 export default class PaginationBoxView extends Component {
   static propTypes = {
@@ -37,6 +38,7 @@ export default class PaginationBoxView extends Component {
     previousLinkClassName: PropTypes.string,
     nextLinkClassName: PropTypes.string,
     disabledClassName: PropTypes.string,
+    disabledLinkClassName: PropTypes.string,
     breakClassName: PropTypes.string,
     breakLinkClassName: PropTypes.string,
     extraAriaContext: PropTypes.string,
@@ -94,8 +96,13 @@ export default class PaginationBoxView extends Component {
   }
 
   componentDidMount() {
-    const { initialPage, disableInitialCallback, extraAriaContext, pageCount } =
-      this.props;
+    const {
+      initialPage,
+      disableInitialCallback,
+      extraAriaContext,
+      pageCount,
+      forcePage,
+    } = this.props;
     // Call the callback with the initialPage item:
     if (typeof initialPage !== 'undefined' && !disableInitialCallback) {
       this.callCallback(initialPage);
@@ -109,16 +116,40 @@ export default class PaginationBoxView extends Component {
 
     if (!Number.isInteger(pageCount)) {
       console.warn(
-        `(react-paginate): The pageCount prop value provided is not an integer (${this.props.pageCount}). Did you forget a Math.ceil()?`
+        `(react-paginate): The pageCount prop value provided is not an integer (${pageCount}). Did you forget a Math.ceil()?`
+      );
+    }
+
+    if (initialPage !== undefined && initialPage > pageCount - 1) {
+      console.warn(
+        `(react-paginate): The initialPage prop provided is greater than the maximum page index from pageCount prop (${initialPage} > ${
+          pageCount - 1
+        }).`
+      );
+    }
+
+    if (forcePage !== undefined && forcePage > pageCount - 1) {
+      console.warn(
+        `(react-paginate): The forcePage prop provided is greater than the maximum page index from pageCount prop (${forcePage} > ${
+          pageCount - 1
+        }).`
       );
     }
   }
 
   componentDidUpdate(prevProps) {
     if (
-      typeof this.props.forcePage !== 'undefined' &&
+      this.props.forcePage !== undefined &&
       this.props.forcePage !== prevProps.forcePage
     ) {
+      if (this.props.forcePage > this.props.pageCount - 1) {
+        console.warn(
+          `(react-paginate): The forcePage prop provided is greater than the maximum page index from pageCount prop (${
+            this.props.forcePage
+          } > ${this.props.pageCount - 1}).`
+        );
+      }
+
       this.setState({ selected: this.props.forcePage });
     }
 
@@ -388,6 +419,7 @@ export default class PaginationBoxView extends Component {
     }
     const {
       disabledClassName,
+      disabledLinkClassName,
       pageCount,
       className,
       containerClassName,
@@ -405,20 +437,31 @@ export default class PaginationBoxView extends Component {
 
     const { selected } = this.state;
 
-    const previousClasses =
-      previousClassName + (selected === 0 ? ` ${disabledClassName}` : '');
-    const nextClasses =
-      nextClassName +
-      (selected === pageCount - 1 ? ` ${disabledClassName}` : '');
+    const isPreviousDisabled = selected === 0;
+    const isNextDisabled = selected === pageCount - 1;
 
-    const previousAriaDisabled = selected === 0 ? 'true' : 'false';
-    const nextAriaDisabled = selected === pageCount - 1 ? 'true' : 'false';
+    const previousClasses = `${classNameIfDefined(previousClassName)}${
+      isPreviousDisabled ? ` ${disabledClassName}` : ''
+    }`;
+    const nextClasses = `${classNameIfDefined(nextClassName)}${
+      isNextDisabled ? ` ${disabledClassName}` : ''
+    }`;
+
+    const previousLinkClasses = `${classNameIfDefined(previousLinkClassName)}${
+      isPreviousDisabled ? ` ${disabledLinkClassName}` : ''
+    }`;
+    const nextLinkClasses = `${classNameIfDefined(nextLinkClassName)}${
+      isNextDisabled ? ` ${disabledLinkClassName}` : ''
+    }`;
+
+    const previousAriaDisabled = isPreviousDisabled ? 'true' : 'false';
+    const nextAriaDisabled = isNextDisabled ? 'true' : 'false';
 
     return (
       <ul className={className || containerClassName}>
         <li className={previousClasses}>
           <a
-            className={previousLinkClassName}
+            className={previousLinkClasses}
             href={this.hrefBuilder(selected - 1)}
             tabIndex="0"
             role="button"
@@ -436,7 +479,7 @@ export default class PaginationBoxView extends Component {
 
         <li className={nextClasses}>
           <a
-            className={nextLinkClassName}
+            className={nextLinkClasses}
             href={this.hrefBuilder(selected + 1)}
             tabIndex="0"
             role="button"